Fixed the issue that InteractivePlotView was not showing as designed. Maybe more elegant solution exists. ----
I made the forcing of a top-level layout onto the component's view optional (and set the value in the component's factory). This works. Perhaps there is a more elegant solution as this feels a slight bit hacky, but acceptable. More important is to make the remainder of the app function properly, if needed, this part can be refactored in some later stage.